Fixed netaxcept rounding problem

+++ com.getshop.client/apps/Netaxept/Netaxept.php
@@ -145,7 +145,7 @@ class Netaxept extends \PaymentApplication implements \Application {
         }
         
         $orderId = $this->order->id;
-        $amount = $this->getApi()->getOrderManager()->getTotalAmount($this->getOrder()) * 100;
+        $amount = (int)($this->getApi()->getOrderManager()->getTotalAmount($this->getOrder()) * 100);
         
         /* var $order \core_ordermanager_data_Order */
         $result = $this->processPayment($amount, $order->paymentTransactionId, $orderId, "CAPTURE");
@@ -253,7 +253,7 @@ class Netaxept extends \PaymentApplication implements \Application {
         $merchid = $this->getConfigurationSetting("merchantid");
         $currency = \ns_9de54ce1_f7a0_4729_b128_b062dc70dcce\ECommerceSettings::fetchCurrencyCode();;
         $orderId = $this->getOrder()->incrementOrderId;
-        $amount = (int)$this->getApi()->getOrderManager()->getTotalAmount($this->getOrder()) * 100;
+        $amount = (int)($this->getApi()->getOrderManager()->getTotalAmount($this->getOrder()) * 100);
 
         $Order = new Order();
         $Order->Amount = $amount; // The amount described as the lowest monetary unit, example: 100,00 NOK is noted as "10000", 9.99 USD is noted as "999".
